added exception for unknow sid

+++ src/database/PrismaSessionStore.ts
@@ -1,9 +1,12 @@
 import * as session from 'express-session';
 import { PrismaClient } from '@prisma/client';
+import { Logger } from '@nestjs/common';
 
 const prisma = new PrismaClient();
 
 export class PrismaSessionStore extends session.Store {
+  private logger = new Logger(PrismaSessionStore.name);
+
   async get(
     sid: string,
     callback: (err?: any, session?: session.SessionData | null) => void,
@@ -51,7 +54,14 @@ export class PrismaSessionStore extends session.Store {
 
   async destroy(sid: string, callback: (err?: any) => void) {
     try {
-      await prisma.session.delete({ where: { id: sid } });
+      const sessionToken = await prisma.session.findUnique({
+        where: { id: sid },
+      });
+      if (sessionToken) {
+        await prisma.session.delete({ where: { id: sid } });
+      } else {
+        this.logger.error(`SID not found in Prisma`);
+      }
       callback();
     } catch (err) {
       callback(err);
